RDKBDEV-1455 : Support Broadband DAC
Reason for change: Support Broadband DAC via distro feature (dac).
Adds DSM recipe. Modifies Dobby recipe so that the thunder plugin
is not a requirement on RDK-B when dac is enabled.
Conditional dac added in package group for RDK-B.
Test Procedure: Enable dac distro feature and ensure dobby and dsm are installed in the rdk-generic-broadband-image build
Risks: Low, components are only affected when the DAC distro feature is enabled

Signed-off-by: Dominic Driver <dominic.driver@consult.red>
Change-Id: Ib40ceaa6982f27ebc086e1ee62c9c599a1a5bf0b