RDKB-48958-Use LGPLv2 version of nettle.
Reason for change: The last LGPLv2 release version of nettle is 2.7.1.
Currently in Kirkstone, we are using nettle version 3.7.3. But it is not
recommended in RDK as it doesn't have LGPLv2 instead it is
LICENSE = LGPL-3.0-or-later | GPL-2.0-or-later.
Nettle 3.7.1 was used because gnutls 3.7.4 requires nettle version above 3.6.
So we need to downgrade both gnutls and nettle so that we can use nettle 2.7.1 version.
Latest neon version (0.30.2) requires latest version of gnutls which required latest version of nettle.
Now, we can remove the configuration of gnutls from neon and have openssl instead of gnutls to avoid such requirements.
Test Procedure: neon should be buildabled. Nettle 2.7.1 version should
be used and not the kirkstone OE version (3.7.3)
Risks: Low

Signed-off-by: Arjun <arjun_daasuramdass@comcast.com>
(cherry picked from commit 586f212a366bddd23ef951d1859d9f12492e6f0b)
Change-Id: I9a63eabe85bf279c36141ff5b5e12f1dbdb6d014