RDKBDEV-3132: Add option to include download_t2_profile.sh
Reason for change: T2 profiles are not downloaded automatically if
web config is not used which is inconvenient for developers.
Test procedures: Configure a clean build with envvar T2_WITHOUT_WEBCONFIG=y.
Build and deploy the image to a device with a valid T2 profile/rule configured in xconf.
Check that is present, and run it "/lib/rdk/download_t2_profile.sh".
Check that the report profile has been added to the data model.
Risks: Low

Change-Id: Ib8fbc6e4366664d1c0e2a7ee5d3ea1be760e2ec5
Signed-off-by: Bryce Young <bryce.young@cognizant.com>