RDKDEV-1112: RDKVREFPLT-2138 [REFPLTV-2290] Update DISTRO VERSION in log
Reason for change: DISTRO_VERSION is still set to "2.0", and used in the logs "Welcome to RDK (A Yocto Project based Distro) 2.0 (dunfell)!", DISTRO_VERSION should be updated to track the version of the RDKV distribution. So, updating the DISTRO_VERSION with the RDK_RELEASE value. Also, updated the RDK_RELEASE to 6.1 from 1.99 in rdk-versions.inc.
Test Procedure: Build and Verified, The log message has been updated to the current RDK_RELEASE.
Risks: None.
Signed-off-by: Sharnetha_SK@comcast.com

Change-Id: I0004f3cb84ff3fbc3163c2f2ab7e809babea8ab0